+++ Changelog.md
@@ -19,6 +19,7 @@
 - torgeternity.chatText.possibilityTooltip
 - torgeternity.sheetLabels.noMin10
 - torgeternity.defeat... (all of them)
+- torgeternity.defeat.chatTitle
 
 
 ## v13.6.0 (AMMO, APPROVED ACTIONS, DSR & UP)